modules/qam-api/requirements.txt
fastapi==0.117.1
uvicorn==0.41.0
psycopg2-binary==2.9.11
pytest==9.0.2
pydantic==2.12.5
